backend/requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
supabase>=2.0.0
openai>=1.0.0
resend>=0.8.0
python-dotenv>=1.0.0
numpy>=1.24.0
flask-cors>=4.0.0
websockets>=12.0
